In Py3, math.floor and math.ceil already return ints.
Remove wrapping int() call.  Mostly to save a line in dates.py...

+++ lib/matplotlib/testing/jpl_units/Epoch.py
@@ -77,7 +77,7 @@ class Epoch(object):
             self._jd = float(jd)
 
             # Resolve seconds down to [ 0, 86400)
-            deltaDays = int(math.floor(self._seconds / 86400.0))
+            deltaDays = math.floor(self._seconds / 86400)
             self._jd += deltaDays
             self._seconds -= deltaDays * 86400.0
